+++ Client/src/Utils/Theme/extractColorObject.js
@@ -1,0 +1,24 @@
+function extractThemeColors(themeType, colorObject) {
+	if (!["light", "dark"].includes(themeType)) {
+		throw new Error('Invalid theme type. Use "light" or "dark".');
+	}
+
+	const extract = (obj) => {
+		if (typeof obj !== "object" || obj === null) {
+			return obj;
+		}
+
+		if ("light" in obj && "dark" in obj) {
+			return obj[themeType]; // Return the value for the specified theme
+		}
+
+		// Reduce the object, keeping only the extracted themeType values
+		return Object.keys(obj).reduce((acc, key) => {
+			acc[key] = extract(obj[key]);
+			return acc;
+		}, {});
+	};
+	return extract(colorObject);
+}
+
+export { extractThemeColors };
